# specs/001-mvp-implementation/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Code Documentation Assistant API
  description: |
    REST API for the Code Documentation Assistant - a RAG-based system that
    ingests codebases and answers natural language questions about code.

    **Key Features:**
    - Upload codebases via ZIP file or GitHub URL
    - Query codebases with natural language
    - Real-time streaming responses
    - Secret detection and redaction
  version: 1.0.0
  contact:
    name: Code Documentation Assistant
  license:
    name: MIT

servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.code-doc-assistant.com
    description: Production

tags:
  - name: codebase
    description: Codebase management operations
  - name: chat
    description: Query and conversation operations
  - name: health
    description: Health check endpoints

paths:
  /api/v1/codebase/upload:
    post:
      tags: [codebase]
      summary: Upload a codebase
      description: |
        Upload a codebase for processing. Accepts either a ZIP/tar.gz file or
        a GitHub repository URL. The system will parse the code, generate
        embeddings, and index it for retrieval.
      operationId: uploadCodebase
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                  description: Human-readable name for the codebase
                  example: "my-python-project"
                description:
                  type: string
                  maxLength: 500
                  description: Optional description
                  example: "A FastAPI web service for user authentication"
                file:
                  type: string
                  format: binary
                  description: ZIP or tar.gz archive (max 100MB)
                repository_url:
                  type: string
                  format: uri
                  pattern: '^https://github\.com/[^/]+/[^/]+'
                  description: GitHub repository URL (alternative to file upload)
                  example: "https://github.com/user/repo"
      responses:
        '202':
          description: Upload accepted, processing started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: File too large (>100MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/codebase:
    get:
      tags: [codebase]
      summary: List all codebases
      description: Retrieve a paginated list of all uploaded codebases
      operationId: listCodebases
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Items per page
      responses:
        '200':
          description: List of codebases
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodebaseListResponse'

  /api/v1/codebase/{codebase_id}:
    get:
      tags: [codebase]
      summary: Get codebase details
      description: Retrieve detailed information about a specific codebase
      operationId: getCodebase
      parameters:
        - $ref: '#/components/parameters/CodebaseId'
      responses:
        '200':
          description: Codebase details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Codebase'
        '404':
          description: Codebase not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [codebase]
      summary: Delete a codebase
      description: Delete a codebase and all associated data
      operationId: deleteCodebase
      parameters:
        - $ref: '#/components/parameters/CodebaseId'
      responses:
        '204':
          description: Codebase deleted
        '404':
          description: Codebase not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/codebase/{codebase_id}/status:
    get:
      tags: [codebase]
      summary: Get ingestion status
      description: Check the processing status of a codebase upload
      operationId: getCodebaseStatus
      parameters:
        - $ref: '#/components/parameters/CodebaseId'
      responses:
        '200':
          description: Current ingestion status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionStatus'
        '404':
          description: Codebase not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/chat:
    post:
      tags: [chat]
      summary: Query a codebase
      description: |
        Ask a natural language question about a codebase. The response is
        streamed via Server-Sent Events (SSE) for real-time delivery.
      operationId: chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Streaming response (text/event-stream)
          content:
            text/event-stream:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    enum: [chunk, sources, done, error]
                    description: Event type
                  content:
                    type: string
                    description: Response content (for type=chunk)
                  sources:
                    type: array
                    description: Source citations (for type=sources)
                    items:
                      $ref: '#/components/schemas/Source'
                  error:
                    type: string
                    description: Error message (for type=error)
              example:
                type: chunk
                content: "The authentication system uses JWT tokens..."
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Codebase not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /health:
    get:
      tags: [health]
      summary: Health check
      description: Basic health check endpoint
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [healthy]
                  version:
                    type: string

  /health/ready:
    get:
      tags: [health]
      summary: Readiness check
      description: Check if the service is ready to accept requests
      operationId: readinessCheck
      responses:
        '200':
          description: Service is ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [ready]
                  dependencies:
                    type: object
                    properties:
                      chromadb:
                        type: string
                        enum: [ok, error]
                      temporal:
                        type: string
                        enum: [ok, error]
        '503':
          description: Service not ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  parameters:
    CodebaseId:
      name: codebase_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique codebase identifier

  schemas:
    UploadResponse:
      type: object
      required: [codebase_id, status, workflow_id]
      properties:
        codebase_id:
          type: string
          format: uuid
          description: Unique identifier for the codebase
        status:
          type: string
          enum: [queued, processing]
          description: Initial ingestion status
        workflow_id:
          type: string
          description: Temporal workflow ID for tracking

    Codebase:
      type: object
      required: [id, name, source_type, status, total_files, size_bytes, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        source_type:
          type: string
          enum: [zip, github_url]
        source_url:
          type: string
          format: uri
          nullable: true
        status:
          type: string
          enum: [queued, processing, completed, failed]
        total_files:
          type: integer
          minimum: 0
        processed_files:
          type: integer
          minimum: 0
        primary_language:
          type: string
          nullable: true
        all_languages:
          type: array
          items:
            type: string
          nullable: true
        size_bytes:
          type: integer
          minimum: 0
        error_message:
          type: string
          nullable: true
        workflow_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CodebaseListResponse:
      type: object
      required: [codebases, total, page]
      properties:
        codebases:
          type: array
          items:
            $ref: '#/components/schemas/Codebase'
        total:
          type: integer
          minimum: 0
          description: Total number of codebases
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100

    IngestionStatus:
      type: object
      required: [codebase_id, status, progress, total_files, processed_files]
      properties:
        codebase_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, processing, completed, failed]
        progress:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Percentage complete (0-100)
        total_files:
          type: integer
          minimum: 0
        processed_files:
          type: integer
          minimum: 0
        current_step:
          type: string
          enum: [validating, cloning, parsing, chunking, embedding, indexing, complete]
          nullable: true
        error:
          type: string
          nullable: true
        secrets_detected:
          type: array
          items:
            type: object
            properties:
              file_path:
                type: string
              secret_count:
                type: integer
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true

    ChatRequest:
      type: object
      required: [codebase_id, query]
      properties:
        codebase_id:
          type: string
          format: uuid
          description: Target codebase to query
        query:
          type: string
          minLength: 1
          maxLength: 1000
          description: Natural language question
          example: "How does the authentication work?"
        session_id:
          type: string
          format: uuid
          description: |
            Session identifier for conversation context.
            If not provided, a new session is created.
        stream:
          type: boolean
          default: true
          description: Whether to stream the response

    Source:
      type: object
      required: [file_path, line_start, line_end]
      properties:
        file_path:
          type: string
          description: Source file path
        line_start:
          type: integer
          minimum: 1
          description: Starting line number
        line_end:
          type: integer
          minimum: 1
          description: Ending line number
        snippet:
          type: string
          description: Code snippet preview
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Relevance confidence score

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Error code
              example: "INVALID_FILE_TYPE"
            message:
              type: string
              description: Human-readable error message
              example: "Only ZIP and tar.gz files are supported"
            details:
              type: object
              additionalProperties: true
              description: Additional error details